Google-Speech-AI/requirements.txt
speech_recognition
pyttsx3
pywin32
pypiwin32
python-engineio
PyAdudio
PocketSphinx
